requirement.txt
Django==2.0.9
django-crispy-forms==1.7.2
django-datatable==0.3.1
python-dateutil==2.7.5
pytz==2018.5
six==1.11.0
